Update Mange MPR and Export Excel

+++ Storage/Storage/DTOs/ExportItemDto.cs
@@ -11,14 +11,14 @@ namespace Storage.DTOs
         public Guid Id { get; set; }
         public DateTime Created { get; set; }
         public string Bill_No { get; set; }
-        public int Sum_Quantity { get; set; }
+        public Int64 Sum_Quantity { get; set; }
     }
 
     internal class ExportItemDetail
     {
         public Guid ExportItemId { get; set; }
         public Guid ItemId { get; set; }
-        public int Qty { get; set; }
+        public Int64 Qty { get; set; }
         public string Note { get; set; }
     }
 }
