+++ Storage/Storage/DTOs/MPRDto.cs
@@ -15,7 +15,7 @@ namespace Storage.DTOs
         public Guid Item_Id { get; set; }
         public string MPR_No { get; set; }
         public string Usage { get; set; }
-        public int Quantity { get; set; }
+        public Int64 Quantity { get; set; }
 
     }
